added mana, stamina and heath costs as well as the display

+++ src/main/java/de/raidcraft/skills/util/HeroUtil.java
@@ -9,10 +9,10 @@ public final class HeroUtil {
 
     private HeroUtil() {}
 
-    public static String createManaBar(int mana, int maxMana) {
+    public static String createManaBar(double mana, double maxMana) {
 
         StringBuilder manaBar = new StringBuilder(String.valueOf(ChatColor.RED) + "[" + ChatColor.BLUE);
-        int percent = (int)(mana / maxMana * 100.0D);
+        int percent = (int)((mana / maxMana) * 100.0);
         int progress = percent / 2;
         for (int i = 0; i < progress; i++) {
             manaBar.append('|');
